move addComment function to utils dir
